Changes the package of :apiTester.
