+++ apiTester/src/commonMain/kotlin/com/revenuecat/purchases/kmp/apitester/PurchasesTransactionExceptionAPI.kt
@@ -1,15 +1,16 @@
-package io.shortway.kobankat.apitester
+package com.revenuecat.purchases.kmp.apitester
 
 import com.revenuecat.purchases.kmp.PurchasesError
 import com.revenuecat.purchases.kmp.PurchasesErrorCode
-import com.revenuecat.purchases.kmp.PurchasesException
+import com.revenuecat.purchases.kmp.PurchasesTransactionException
 
 @Suppress("unused", "UNUSED_VARIABLE")
-private class PurchasesExceptionAPI {
-    fun check(exception: PurchasesException) {
+private class PurchasesTransactionExceptionAPI {
+    fun check(exception: PurchasesTransactionException) {
         val underlyingErrorMessage: String? = exception.underlyingErrorMessage
         val message: String = exception.message
         val code: PurchasesErrorCode = exception.code
         val error: PurchasesError = exception.error
+        val userCancelled: Boolean = exception.userCancelled
     }
 }
